Move WT_VERSION into its own file
