+++ src/include/wt_internal.h
@@ -440,6 +440,7 @@ typedef uint64_t wt_timestamp_t;
 #include "verbose.h"
 
 #include "session.h" /* required by connection.h */
+#include "version.h" /* required by connection.h */
 #include "connection.h"
 
 #include "extern.h"
